+++ src/Sulu/Bundle/MediaBundle/Tests/Functional/Entity/CollectionRepositoryTest.php
@@ -16,6 +16,8 @@ use Sulu\Bundle\MediaBundle\Entity\Collection;
 use Sulu\Bundle\MediaBundle\Entity\CollectionMeta;
 use Sulu\Bundle\MediaBundle\Entity\CollectionRepository;
 use Sulu\Bundle\MediaBundle\Entity\CollectionType;
+use Sulu\Bundle\MediaBundle\Entity\Media;
+use Sulu\Bundle\MediaBundle\Entity\MediaType;
 use Sulu\Bundle\TestBundle\Testing\SuluTestCase;
 use Sulu\Component\Media\SystemCollections\SystemCollectionManagerInterface;
 
@@ -27,30 +29,35 @@ class CollectionRepositoryTest extends SuluTestCase
     private $em;
 
     /**
+     * @var MediaType
+     */
+    private $mediaType;
+
+    /**
      * @var array
      */
     private $collectionData = [
-        ['1', null, false],
-        ['2', 0, false],
-        ['3', 1, false],
-        ['4', 2, false],
-        ['5', 3, false],
-        ['6', 3, false],
-        ['7', 5, false],
-        ['8', 5, false],
-        ['9', 2, false],
-        ['10', 1, false],
-        ['11', 9, false],
-        ['12', 9, false],
-        ['13', 0, false],
-        ['14', 12, false],
-        ['15', 12, false],
-        ['16', null, true],
-        ['17', 15, true],
-        ['18', 15, true],
-        ['19', 17, true],
-        ['20', 17, true],
-        ['21', null, false],
+        ['1', null, false, 0],
+        ['2', 0, false, 0],
+        ['3', 1, false, 0],
+        ['4', 2, false, 4],
+        ['5', 3, false, 0],
+        ['6', 3, false, 0],
+        ['7', 5, false, 0],
+        ['8', 5, false, 0],
+        ['9', 2, false, 0],
+        ['10', 1, false, 0],
+        ['11', 9, false, 0],
+        ['12', 9, false, 0],
+        ['13', 0, false, 0],
+        ['14', 12, false, 0],
+        ['15', 12, false, 0],
+        ['16', null, true, 0],
+        ['17', 15, true, 0],
+        ['18', 15, true, 0],
+        ['19', 17, true, 0],
+        ['20', 17, true, 0],
+        ['21', null, false, 0],
     ];
 
     /**
@@ -82,11 +89,17 @@ class CollectionRepositoryTest extends SuluTestCase
         $systemCollectionType->setKey(SystemCollectionManagerInterface::COLLECTION_TYPE);
         $this->em->persist($systemCollectionType);
 
+        $this->mediaType = new MediaType();
+        $this->mediaType->setName('image');
+        $this->mediaType->setDescription('This is an image');
+        $this->em->persist($this->mediaType);
+
         foreach ($this->collectionData as $collection) {
             $this->collections[] = $this->createCollection(
                 $collection[0],
                 $collection[1],
-                $collection[2] ? $systemCollectionType : $defaultCollectionType
+                $collection[2] ? $systemCollectionType : $defaultCollectionType,
+                $collection[3]
             );
         }
         $this->em->flush();
@@ -97,7 +110,7 @@ class CollectionRepositoryTest extends SuluTestCase
         $this->em->flush();
     }
 
-    private function createCollection($name, $parent, $collectionType)
+    private function createCollection($name, $parent, $collectionType, $numberMedia)
     {
         $collection = new Collection();
         $collectionMeta = new CollectionMeta();
@@ -112,10 +125,24 @@ class CollectionRepositoryTest extends SuluTestCase
             $this->collections[$parent]->addChildren($collection);
         }
 
+        $this->addMedia($collection, $numberMedia);
+
         $this->em->persist($collection);
         $this->em->persist($collectionMeta);
 
         return $collection;
+    }
+
+    private function addMedia(Collection $collection, $numberMedia)
+    {
+        for ($i = 0; $i < $numberMedia; ++$i) {
+            $media = new Media();
+            $media->setCollection($collection);
+            $media->setType($this->mediaType);
+            $collection->addMedia($media);
+
+            $this->em->persist($media);
+        }
     }
 
     public function provideTreeData()
@@ -160,6 +187,16 @@ class CollectionRepositoryTest extends SuluTestCase
     public function testCount()
     {
         $this->assertEquals(16, $this->collectionRepository->count(5, ['systemCollections' => false]));
+    }
+
+    public function testCountMedia()
+    {
+        $this->assertEquals(4, $this->collectionRepository->countMedia($this->collections[3]));
+    }
+
+    public function testCountMediaWithEmptyCollection()
+    {
+        $this->assertEquals(0, $this->collectionRepository->countMedia($this->collections[0]));
     }
 
     public function testCountWithoutFilters()